Separate out inspect flag end-to-end emulator test.
Unfortunatley, shutting down/start up emulator is too flakey.
