+++ .github/workflows/node-test.yml
@@ -83,6 +83,7 @@ jobs:
           - npm run test:import-export
           - npm run test:extensions-emulator
           - npm run test:triggers-end-to-end
+          - npm run test:triggers-end-to-end:inspect
           - npm run test:storage-deploy
           - npm run test:storage-emulator-integration
     steps:
@@ -137,6 +138,7 @@ jobs:
           # - npm run test:import-export # Fails becuase port 4000 is taken after first run - hub not shhutting down?
           # - npm run test:extensions-emulator # Fails due to cannot find module sharp (not waiting for npm install?)
           - npm run test:triggers-end-to-end
+          - npm run test:triggers-end-to-end:inspect
           - npm run test:storage-deploy
           # - npm run test:storage-emulator-integration
     steps:
